// admin_panel/app/volunteers/types.ts
//volunteers/types
export interface Volunteers {
    id: string;
    first_name: string;
    last_name: string;
    email: string;
    phone_number: string;
    preferred_county: string;
    volunteer_type: string;
    skills: string;
    time_commitment: string;
    preferredTime: string;
    status: string;
    date: string;
}